Added better documentation and moved all GRPO experimental files to experimental/grpo
