+++ nemo_aligner/experimental/grpo/data/builders.py
@@ -1,7 +1,7 @@
 from functools import partial
 
 from nemo_aligner.data.nlp.builders import build_train_valid_test_datasets
-from nemo_aligner.experimental.data.datasets import AllTaskDataset, environment_collate_with_batch_max_sequence_length
+from nemo_aligner.experimental.grpo.data.datasets import AllTaskDataset, environment_collate_with_batch_max_sequence_length
 
 build_train_valid_test_task_datasets = partial(build_train_valid_test_datasets, AllTaskDataset)
