+++ nemo_aligner/experimental/grpo/experience/environments/math_environment.py
@@ -5,8 +5,8 @@ import torch
 
 from nemo_aligner.utils import parallel_state
 from nemo_aligner.utils.utils import masked_mean
-from nemo_aligner.experimental.experience.interfaces import EnvironmentInterface
-from nemo_aligner.experimental.experience.environments.metrics import calculate_pass_rate_per_prompt
+from nemo_aligner.experimental.grpo.experience.interfaces import EnvironmentInterface
+from nemo_aligner.experimental.grpo.experience.environments.metrics import calculate_pass_rate_per_prompt
 from nemo_aligner.servers.http_communicator import FlaskCommunicator
 
 class MathEnvironment(EnvironmentInterface):
